[inductor] Disable fp contraction and add precise division option
Fixes #101039, Fixes #122179

In the issue we have something like:
```python
(x / 1e-6) + (y / 1e-6)
```
which triton turns into
```python
fma(x, 1e6, y * 1e6)
```

Where we have one division calculated to float32 precision,
and the other calculated to infinite precision as part of the `fma`.
In the issue this leads to a bad normalization in softmax.

Here we disable `fma` fusion in the trion compiler.

ghstack-source-id: aa9f68f10abf93a6f129468560e4f36660d5949d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115435