Fix mask propagation in the presence of where
Before, when calling ops.where, masks were not properly propagated. We
now restrict the optimisation to `ops.masked`, which I think it was what
the original code intended to do.

I'm not 100% sure that even in the masked case this code is not
introducing some bugs, but this is a strict improvement over the
previous state.

ghstack-source-id: bf1df215a400d8578023dfc0dff2ae28f961d602
Pull Request resolved: https://github.com/pytorch/pytorch/pull/125574